Added a logical to avoid adding hyper viscosity to the 3 cells next to the boundary.
Hyperviscosity can give problems when it's taken across the boundary since it's taking a 6th-derivative across it and we don't have a boundary condition that is 6th order smooth (do we?), hence some weird modes grow exponentially. Border driving helps, but the best solution short of a 6th order smooth boundary condition is to *not* add hyperviscosity in the first and last three cells.

To use this patch, add lhyper_avoid_boundary in the run_pars of run.in

Thanks Colin for noticing this.