Switch Fourier Fargo (Fargo + SAFI) as default.
Hm... how to combine both names?

MAJAL? The Masset algorithm, Johansified and Lyrified?

Hm, a "Taj" before would definitely make it popular in India. Lemme see.. TAJ can stand for... Tom and Jerry.. TAJ MAJAL: Tom and Jerry: the Masset Algorithm, Johansified and Lyrified. Tom and Jerry here is a metonimy for any duet (and allows for interesting puns when giving talks). Tom may also mean the large shift, and Jerry the residual. That obvioulsy begs the questions of what is the Fourier transform of Schrodinger's cat. Meow.

Though swapping the H for a J retaining its sound as voiceless velar fricative (IPA x) betrays the author's Ibero-American influences and, while placing him on the correct margin of the Ebro, erroneously assigns him to the Hispanic side of the peninsula. Perhaps a more Lusitan approach is warranted, in which case the algorithm should be dubbed ....

FAROFA?

Fast Advection for Rotating objects using Fourier... Auspiciouness?

Farofa is a Brazilian dish.