Bug caught.
The advective derivative in cylindrical coordinates was wrong for ugradA and hgradA

The term should be

(A.nabla)B = [(A.nabla)B]_0 - Aphi*Bphi/r \hat{r} + Aphi*Brad/r \hat{phi}

It was coded with Arad*Bphi/r in the last term.

It never affected p%ugu, but p%uga in the advective gauge was wrong, as any (h.grad)f in cylindrical coordinates.