Public variables in magnetic.h?
A new dummy but public module variable lcalc_aamean needs to be defined
in bfield.f90 because of the change in magnetic.h in commit dc79b57d.
I though most of us agreed that it is not a good practice to make module
variables public, but currently there are many at least in magnetic.h.
Anyway...