changed typos
check if the cylinder is placed fully in the domain. fatal error if otherwise

the o-grid related variables ivar1_part and ivar2_part are now namelist inputs instead of being hardcoded. These variables decide which data of the o-grid has to be communicated to the particles that are on the current processor space-wise, but are stored on another processor. the default values are 1 and 4, which corresponds to a communication with iux to irho.