Fix a bug introduced in commit c2e3fbfd.
When using Cartesian coordinates, the code died at pencil checks.

The reason was the results calculated by div_mn() and div_other() are
not exactly the same.  The same happens to curl_mn() and curl_other().

Current workaround is to disable the selection between div_mn() and
div_other() and between curl_mn() and curl_other() for p%diva and p%bb,
respectively, when doing pencil checks.

This difference due to the choice of calculations, however, may be
worrying.