Replace some usage of sed with POSIX compliance.
See the comments at comment 1ee6da3.

It turns out that many advanced features in different implementations of
sed are not cross platform.  The only way to guarantee portability is to
follow the POSIX standard.  See, e.g.,

https://riptutorial.com/sed/topic/9436/bsd-macos-sed-vs--gnu-sed-vs--the-posix-sed-specification

In this commit, a Makefile function sed_follow_link is implemented to
emulate -i and --follow-symlinks options.