Revise migrate_particles() of particles_mpicomm.
Follows commit c70651a and commit e37fd12.

Before comparing the position of a particle with the process boundary,
shift both against xyz0 so that the round-off errors are consistent with
how ineargrid is computed in subroutine map_nearest_grid() of
src/particles_map.

TODO: It remains a better programming practice that both subroutines use
the same method to compute which cell each particle belongs to (and
hence if a particle leaves a process).