Fix a bug in particle species assignment.
The multiplication of npar_species and npar can overflow when either
npar_species or npar is too large.  Also, there are too many places that
compute the species of a particle given its ipar.

In this commit, the assignment of a particle to its species is
centralized in one function assign_species() under module Particles_sub.
All occurrences of this computation in particles_dust.f90 are replaced
by a call to assign_species().

TODO(): There are many such occurrences in other particles module, which
also need to be replaced.