Use plt.pause instead of sleep
sleep just pauses the entire thread (meaning matplotlib also will not be
able to update the animation in that gap) On the other hand, plt.pause
allows matplotlib to update the animation during the pause interval.

Without this change, only a blank window is shown (on Qt/X11/Linux)
until the loop ends. With this change, the animation works correctly.