Added patch to call sum_par_name from every processor (even those that do not have any particle).
sum_par_name was being called from the do k=1,npar_loc loop for diagnostic potselfpm. Only some processors were
calling it, and the code was hanging as a result. The patch adds a call, after the particle loop, to be activated
if the processor does not have particles. Ideally it should be a single call, after the k loop.

Thanks Chao Chin for the patch.