Added to the discuss group but adding here too to register it in the log
Hi all

tl;dr: Hyperdiffusion was removed from the code in October 2018. All runs that use hyperdiffusion since then are BROKEN. That this was not caught in 3 yrs exposes a crack in the development style of the code that we need to fix to ensure nothing of the sort ever happens again.

I found a massive problem that has been going on since 2018.

The code is not adding hyperdiffusion.

Looking in density.f90, fdiff is added, and then hyperdiffusion is calculated.

But the contribution of hyperdiffusion to fdiff is never added to df.

I’ve traced in the logs what happened.

commit 995aa8b59838ce18618081bcdeac2120c5d6cf65
Author: Frederick Gent <fred.g...@gmail.com>
Date:  Mon Oct 22 12:53:33 2018 +0300  Moved shock diffusion and massdiff_fix ahead of hyperdiffusion

This commit moved all hyperdiffusion calculations to after fdiff is added to df. Since this commit, hyperdiffusion is calculated but ignored.

This immediately broke the auto-test on a number of samples. Not all samples that use hyperdiffusion broke, because hyper only affects noise at the grid scale. The samples that broke are the ones below

./cylindrical-globaldisk-dzone
./baroclinic
./MRI-turbulence_hyper
./2d-tests/debris-photoeletric-streaming
./2d-tests/planet-disk-fargo
./2d-tests/Kelvin-Helmholtz
./2d-tests/selfgravdisk-logspirals

Looking at the logs, the reference.out files were updated the day after the crippling edit
./cylindrical-globaldisk-dzone
commit 95f51f530572a1c4be8fefe04e1e287604896790
Author: Philippe Bourdin <11717935+...@users.noreply.github.com>
Date: Tue Oct 23 17:40:34 2018 +0200
Adjusted precision of diagnostic output.

./baroclinic
commit acd8c57f3d17878ed8ede74da7dd08879f6f64cd
Author: Philippe Bourdin <11717935+...@users.noreply.github.com>
Date: Tue Oct 23 16:53:03 2018 +0200
Removed quantities that are dominated by noise and lowered precision where necessary.

./MRI-turbulence_hyper
commit 5c11ed83e7f55eabb287bc592ec2ea7a438e0c15
Author: Philippe Bourdin <11717935+...@users.noreply.github.com>
Date: Tue Oct 23 16:39:00 2018 +0200
Reference output missing in the last commit.

./2d-tests/debris-photoeletric-streaming
commit fa1b94f2d28e84cf5ad04d7f6362ccdf26d3195a
Author: Philippe Bourdin <11717935+...@users.noreply.github.com>
Date: Tue Oct 23 17:06:06 2018 +0200
Fixed precision of diagnostic output for this auto-test.

./2d-tests/planet-disk-fargo
commit a9c557298125d90f0d31fcd8aedbbd2247112c45
Author: Philippe Bourdin <11717935+...@users.noreply.github.com>
Date: Tue Oct 23 14:53:10 2018 +0200
The precision in those quantities changed here is simply not sufficient to compare all these digits.

./2d-tests/Kelvin-Helmholtz
commit 06be507a7e264043cfddfc3270ebf3ce1265829f
Author: Philippe Bourdin <11717935+...@users.noreply.github.com>
Date: Tue Oct 23 17:18:49 2018 +0200
Ajusted the precision according to the reference data.

./2d-tests/selfgravdisk-logspirals
commit d0419696afab17bc2652aab609135ac92863ae29
Author: Philippe Bourdin <11717935+...@users.noreply.github.com>
Date: Sat Oct 20 06:01:42 2018 +0200
Removed unreliable quantities that strongly depend on the compiler version I use.

That is, the fix actually hid the problem by lowering resolution and removing the quantities that did accuse the serious flaw that got introduced. To be fair, a fix to the auto-test was needed because some variables tested had underflow. But this was a full-scale red alert in response to the murder of hyperdiffusion, yet the aggressive fix removed the only guard-rails that we had against it. With the auto-test quiet, the error went unnoticed.

The error was corrected in the March 2019 commit

52e57714b33642005a636396b4f930e143872077
Author: Frederick Gent <fred.g...@gmail.com>
Date:  Tue Mar 12 13:01:07 2019 +0200  Corrected order of diffusivities and option to load old rho to run lnrho

Yet, the auto-test again yelled bloody murder. I reverted the commit.

commit 42349a5dfb408e98dcc319ce677d18b7340eeb1b
Author: wlyra <wladim...@gmail.com>
Date:  Sat Mar 30 14:30:58 2019 -0700  Undoing commit that broke the auto-test :  https://github.com/pencil-code/pencil-code/commit/52e57714b33642005a636396b4f930e143872077  The problem was in some change to the density.f90, could not trace it exactly since the commit has a lot of stuff. Fred, please check.
And no one has touched it since.

I caught this by spending this whole weekend tracing why a simulation was crashing. I eventually pinpointed that it was the hyperdiffusion, after ruling out a number of other possibilities. I isolated the problem by setting a Nyquist signal in density, with zero velocity, and comparing the resulting amplitude after one Euler timestep for the various hyperdiffusion formulations.

To my surprise, there was no change in density.

Tracing the code, I saw that fdiff was never added to df.

This is unthinkable: an error of this magnitude going unnoticed for this long, with samples of different owners being tweaked.

It's a tricky error to catch since it happens at the grid scale, and not all users use hyperdiffusion. We should have a sample that catches all sorts of possible hyper.

Yet, we need to discuss in future meetings a system beyond the auto-test to not allow this to happen again.

Best wishes,
Wlad