Fix u2mz diagnostic
Apparently whoever added the u2mz diagnostic forgot to add it to the
lreset block. This means that pencil will wrongly comment out the entry
in the xyaver.in file (but write out the average anyway in
xyaverages.dat). Unsuprisingly, this causes a lot of problems while
reading the averages later.