Python: read var: support nonequidistant grids for magic functions
Earlier, calculation of magic functions (bb, vort, jj) supported only
equidistant grids.

NOTE: I am not sure whether we need to support reading varfiles in the absence
of a grid.dat at all. As of this commit, it is supported just to make sure a test
(test_read_var) does not fail. If we decide not to support that use-case, we can
just add a grid.dat to `tests/input/serial-1/proc0` and simplify the
code here.