entropy.f90: Mark a BUG in diagnostic
fradtop is not initialized properly (this was caught by pencil_check).
The first call of surf_mn_name for this diagnostic will be when
lfirstpoint=.false., so surf_mn_name will not initialize the diagnostic
properly.

I'm just leaving a comment marking the bug since I am not sure what the
cleanest way to fix it would be. Perhaps it is not a good idea to depend
on lfirstpoint in surf_mn_name.