Python: modify is_int test
If the input string contained '.', treat it as a float regardless of
whether it can be losslessly represented as an integer.

This is useful during parameter study. E.g. if the template sim had
'Omega=0.', the older version would assume Omega is an int, and then
throw an error when one tries to change the value to, say, 1e-1.