entropy.f90: correctly set cs2top when cs2cool2/=0
cs2cool2 is used in the cooling profile 'two-layer'. When these profiles
are used, the cs2top should be a weighted sum of cs2cool and cs2cool2.
Earlier code would unconditionally set it to cs2cool.

cs2cool is also used in the profile 'shell3', but I am not sure what to
do there.