forcing_chandra_kendall: restore commented line
It was commented in commit 9bc92e347dde18476e823a7fcda6b7afe918f0fb with
no explanation. After that, the variable rz (used in the next line) was
uninitialized.

I am restoring this line assuming it was commented by accident. Note
that
uninitialized variables being 0 (compiler-dependent) made the code still
seem to work, but we should not depend on this behaviour.