diagnostics.f90: yzsum_mn_name_x_mpar: fix normalization for non-Cartesian coordinates
The fact that the line elements for the y and z coordinates in
cylindrical/spherical coordinates are x-dependent was accounted for
while summing the values, but not while normalizing the final result to
get the average. I have now ensured that both are handled in a
consistent way to give the correct result.