Python: Power._read_power2d: use array rather than list
It turns out that using an intermediate list is very expensive in terms
of RAM (each element of the list is a pointer to a memory address,
consuming 8 bytes of memory even before we account for the memory
consumed by the list entries themselves).

The alternative is to start with a numpy array, and resize it each time
a new line is read. Naively, one expects this to be very expensive in
terms of CPU time, but it turns out to only be slightly slower than just
maintaining a list.

Behold the following benchmarks in an IPython session:
```
In [1]: import numpy as np

In [2]: def fun_1(n): #old approach
   ...:     l = []
   ...:     for i in range(int(n)):
   ...:         l.append(i)
   ...:     return np.array(l, dtype=np.single)
   ...:

In [4]: def fun_3(n, block_size=100): #new approach
   ...:     l = np.array([], dtype=np.single)
   ...:     for i in range(int(n/block_size)):
   ...:         l.resize((len(l)+block_size,))
   ...:         for i in range(block_size):
   ...:             l[-1-i] = i
   ...:     return l
   ...:

In [5]: %timeit fun_1(1e5)
9.55 ms ± 45.3 µs per loop (mean ± std. dev. of 7 runs, 100 loops each)

In [7]: %timeit fun_3(1e5)
11.6 ms ± 118 µs per loop (mean ± std. dev. of 7 runs, 100 loops each)
```