manual: fix mistake in expression for the forcing_hel
Axel (or anyone else familiar with forcing_hel), please check.

The updated expression satisfies i k x f = |k|f as mentioned in the
manual, and also reduces to the mentioned form for sigma=0.

This was pointed out by Rajesh Mondal who says that the updated
expression here is what is actually being calculated in the code.