start.f90: move init_eos into the init_loops loop
The issue I am trying to deal with is: in eos_idealgas_vapor, cp is not
a constant (calculated based on the value of another variable), and so
is stored in the f-array. If any initial condition wants to use this
non-constant cp, it needs to have the cp already present in the f-array.
The only idea I have right now is to populate the f array in init_eos,
and then use init_loops=2 so that the correct value of cp will be picked
up on the second run.

This is of course a dirty hack, so better suggestions are welcome.