eos.h: add eoscalc_point_f to interface
Calculates EOS variables given the f-array at a single spatial point.

eoscalc_point seems impossible to implement for eos_idealgas_vapor,
since the temperature etc. depend on components of the f-array that are
not present in eosvar{1,2}

For EOS modules which implement eoscalc_point, eoscalc_point_f is just a
wrapper around that.