Python: tests/manual.py: update reference values for test_read_var
Necessitated by my recent replacement of eos_idealgas.f90/bc_ss_flux by
noeos.f90/bc_ss_flux. Although my change did not cause the auto-test to
fail, this Python test validates values to a higher precision. For
example, the test `np.mean(f[1, :, :, :])` was being tested to a
precision of 1e-17 (!).

I have checked that if the grid size is changed to 64^3 (from 32^3, and
reference values are changed accordingly), the replacement of the
boundary condition does not cause the Python test to fail (excepting the
one that was being tested to precision 1e-17, which did not converge
even with a 128^3 grid). Nevertheless, I have not changed the grid size
of the sample.