io_hdf5: read_persist_*_0D: do not use uninitialized variable when reading fails
In io_dist, the behaviour is that if reading fails, the pre-existing
value of the variable 'value' is preserved. On the other hand,
io_hdf5/read_persist_*_0D would set 'value' to an uninitialized variable
if reading were to fail.

This commit changes the behaviour of the implementations in io_hdf5 to
match the (saner) behaviour of the ones in io_dist.

I only noticed this because I recently started passing -finit-real=nan
to gfortran.