hydro: fix out of bounds access in diagnostics
The T*m diagnostics access the f-array directly, and so need to be
guarded behind if conditions. Ideally these quantities would be pencils,
but I am just making a quick fix for now.

There seems to be a bug in the T0z2m diagnostic. Axel, could you please
check? (It looks like you coded it)

Hooray for -fcheck=bounds .

The bug was exposed by Matthias' recent simplifications in commit
27836f131391fe973f3bf46026b260d82bbff14a .