samples/helical-MHDTurb/tests: partially update reference data
Account for recent changes wherein dt=4e-2 was added to run.in, dspec
and d2davg were changed, and REAL_PRECISION was changed to double.

The reference data in this commit allow the tests to pass when the rest
of the code is at commit 5456591699ffcae5875daa429022fc60f6e5f38e . As
detailed in [1], I suspect a bug has crept in after that; let's wait
until we get to the bottom of that before updating the reference data
further.

Needless to say, it's a pain to have to track down all these changes in
order to convince oneself that the changed results are not due to
changes in the code itself. I think changes to the input parameters of
auto-tested samples should be avoided unless absolutely necessary.

Historical notes on what changed in which commit:

Commit [A] changed d2davg, dspec, lspec_start
Updating
- read_data.ref
- read_slices.ref
- read_yaver.ref
- read_zaver.ref
to the values as of commit [A] makes tests pass till (not including)
commit [B], where dt=4e-2 was added to run.in (!!).
In the next commit [C], the precision was changed to double.
With updated *.ref from commit [C], the script tests also pass [D],
which removed lspec_start=T from run.in.
In between [D] and [E], gputestv6 seems to have been merged into master.
Due to some change in this monster merge, 1D and 2D averages are no
longer output at t=0, causing the script tests to fail again.

References/commit IDs:
- [1]  :
https://groups.google.com/g/pencil-code-discuss/c/6WJkJqpJBqQ/m/xk6KGAxbAAAJ

- [A]: 0b13f7b05b87a1f8cf51f20adfcdfc34d24ccdb9
- [B]: d77a80c7ffdeca8128c2d41fb3c42e721dbb8071
- [C]: d922e9bed3b9ccf36b0013448d95755338cd4215
		888032d981e9b7a1f026ad9bfa16066cd933aefa: good
		1dfb260a97e2a682602295be52a39b699c17da09: bad
- [D]: 5456591699ffcae5875daa429022fc60f6e5f38e : good
	099e0caa724ebeac2735cac0fb075b6afd4afc64: good
	c762afe3511af3befaa16d0e09fe42437102e6e3: does not compile
(merge master into gputestv6)
- [E]: b08909028a988b208421d5215dc9415fcfb05aa4