This test output is the same as after check-in 686de31, that fixed hyperdiffusion. The sample already had lmassdiff_fix = .true. so no fix to it to deal with the shock mass diffusion was needed.
Yet, the test itself was changed, specifically, the initial condition in start.in, on commit c1262a4,
was changed from ‘linear-zero’ to ‘solid-body’, apparently to circumvent the negative rho (which is
ok for this sample since this is shallow water where rho is not density but deviation from reference
height -- we had a switch to allow negative rho; reference publication https://ui.adsabs.harvard.edu/abs/2022PSJ.....3..166H/abstract).

Again, this was done without asking the maintainers.

Since this commit, May 25, 2021, this has been A COMPLETELY DIFFERENT TEST THAN THE MAINTAINERS INTENDED.

WTF!!!