Add sample to test power_xy subroutine
After the gputestv6 merge, setups that use power_xy suffer from an
out-of-bounds array access.

I am now adding a sample to demonstrate this. Reference data are from a
branch of mine that does not include the changes in gputestv6. Make sure
to compile it with the bounds checking options enabled in your compiler
(see config/compilers/extensions/GNU-GCC_debug.conf ).

I haven't had time to try to fix the issue yet. Hopefully it won't turn
out to be too difficult to fix.