Add script pc_isolated-test
Clones pencil-code into an isolated working copy and runs the auto-tests
from that.

Since tests can run for hours, it is inconvenient to have to keep your
local copy of the pencil-code unchanged till the test finishes. This
script can be used to make a copy of your current local version of
pencil and run tests in the copy, allowing you to continue working on
your local copy while the tests run.

A usage example is in the comments at the beginning of the file