Add sample: HDF5 version of helical-MHDTurb
Mainly so that we can also test writing and reading (in Python) the HDF5
averages.

The current reference data are the same as in the standard (nohdf5,
nompicomm) sample. Note that the 2D average reading tests account for
some quirks of the Python module in order to avoid changing the
reference data.