helical-MHDTurb: Revert "Updated reference data with correct timestamps"
This reverts commit e58fdf98176cf975253305690ea8e3bdf6118077

Later discussion made it clear that after the gputestv6 merge, the 1D
and 2D averages are written with timestamp t+dt, but with values
apparently corresponding to t. My understanding (from what Fred said
during the Office Hours) is that this update of the reference data
was not intentional. The issue is still not fixed, but at least let's
restore the reference data so that the test fails (as it should).

First raised in https://groups.google.com/g/pencil-code-discuss/c/6WJkJqpJBqQ/m/xk6KGAxbAAAJ