Fix timestamps of 1D and 2D averages being t+dt rather than t
The issue is that gen_output (which calls write_{1,2}daverages) is
called after the time_step subroutine increments t to t+dt. However, the
averages are actually calculated using the fields at time t.

This commit changes write_{1,2}daverages to use tdiagnos for the
timestamp, rather than t.

See
https://groups.google.com/g/pencil-code-discuss/c/6WJkJqpJBqQ/m/xk6KGAxbAAAJ

Matthias/Fred/Touko, can one of you check if this affects the
functioning of the GPU version of the code? I don't have a working GPU
setup to test this myself.