equ.f90: test_methods/testfield_x.f90: added: +! CPARAM logical, parameter :: ltestfield_x = .true. +! CPARAM logical, parameter :: ltestfield_z = .false. +! CPARAM logical, parameter :: ltestfield_xy = .false. +! CPARAM logical, parameter :: ltestfield_xz  = .false.
test_methods/testfield_z.f90: needed to comment out
+        !allocate(p(1),lpenc_loc(npencils))
because now it comes in as an argument

Hi Matthias,
In equ.f90, I undid the:

      !if (ltestfield)      call testfield_after_boundary(f)
!AB: quick fix
      if (ltestfield)      call testfield_after_boundary(f,p)

to now include p, so that testfield_x works.
However, in testfield_z, the p was handled differently
and now the auto-test for testfield_z doesn't work.
I check it in anyway in the hope that we can fix it quickly.
I also check in the 3 diff filex, and the critical places
can be found by searching for "quick fix".
I thought you might know about the p problem, because
I think you worked with this last time, and now
Gustav (cc) needs that.
Axel