Python: implement lazy loading of submodules
Since some of the submodules (which may not even be used by many people)
take a lot of time to import, we now only initialize them when needed
(rather than initializing all of them when the pencil module is
imported).

This adds lazy_loader (Python package) as a dependency. I think this is
fine, since that package is even available in the Ubuntu repositories
(python3-lazy-loader).